trace raft
