+++ Cargo.toml
@@ -161,8 +161,6 @@ url = "2"
 uuid = { version = "0.8.1", features = ["serde", "v4"] }
 vlog = "0.1.4"
 walkdir = "2"
-minitrace = { git = "https://github.com/pingcap-incubator/minitrace-rust.git", branch = "master", features = ["jaeger"] }
-
 
 [dependencies.prometheus-static-metric]
 git = "https://github.com/tikv/rust-prometheus.git"
@@ -188,7 +186,8 @@ fail = { git = "https://github.com/tikv/fail-rs.git", rev = "2cf1175a1a5cc2c70bd
 prometheus = { git = "https://github.com/tikv/rust-prometheus.git", rev = "fd122caa03de8e7e5e4fae9372583aebf19e39f6" }
 
 [patch.'https://github.com/pingcap/tipb.git']
-tipb = { git = "https://github.com/zhongzc/tipb.git", branch = "tracing", default-features = false }
+tipb = { path = "../tipb", default-features = false }
+#tipb = { git = "https://github.com/zhongzc/tipb.git", branch = "tracing", default-features = false }
 
 [target.'cfg(target_os = "linux")'.dependencies]
 procinfo = { git = "https://github.com/tikv/procinfo-rs", rev = "5125fc1a69496b73b26b3c08b6e8afc3c665a56e" }
