+++ components/tikv_util/Cargo.toml
@@ -52,6 +52,8 @@ tokio-sync = "0.1.7"
 tokio-timer = "0.2"
 url = "2"
 kvproto = { git = "https://github.com/pingcap/kvproto.git", default-features = false }
+minitrace = { git = "https://github.com/pingcap-incubator/minitrace-rust.git", branch = "master", features = ["jaeger"] }
+protobuf = "2.8"
 
 [target.'cfg(target_os = "linux")'.dependencies]
 procinfo = { git = "https://github.com/tikv/procinfo-rs", rev = "5125fc1a69496b73b26b3c08b6e8afc3c665a56e" }
